I made the way I set the speeds of everything more elegant
Originally when I was testing, I just calculated a speed based on a random angle I chose, and then set the launcher to that speed. However this is not optimal. its much more elegant if you put the velocity in the feasible range, and then set the angle based on the velocity it ends up at.

While I was at it, I tried to clean up the teleop code a little.